+++ NEWS.md
@@ -12,6 +12,7 @@
 - Fixed `nnf_normalize` and `x$norm()` as they were not able to be called (#409)
 - Made `torch_one_hot` and `nnf_one_hot` use 1-based indexing. (#410)
 - Updated to LibTorch 1.7.1 (#412)
+- Allow listing all nested modules in a `nn_module` (#417)
 
 # torch 0.2.0